+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Pearl changelog
 
 #### Changed
 - Save delete records in index (#207)
+- Use bytes for arguments (#191)
 
 
 #### Fixed